stable/concourse: allow taking over the `concourse-work-dir` volume
This allows using a user-defined volume spec (set in
`worker.additionalVolumes` with `name: concourse-work-dir`) instead of
the default `emptyDir` when `persistence.enabled: false` is set.

It changes the behaviour in a backward-compatible way: previously using
`name: concourse-work-dir` in any `worker.additionalVolumes` resulted in
an error (two volumes defined with the same name), now instead of the
error it will use that additionalVolume spec to back the
`concourse-work-dir` volume mount (given that `persistence.enabled` is
false).

Signed-off-by: Norbert Buchmueller <norbert.buchmueller@optiopay.com>